rename commit funds components and move some to separate files
