+++ app/components/commitfunds/CommitUnknownUser.scss
@@ -1,0 +1,6 @@
+@import '../../../common/sass/variables';
+
+.initial-link {
+  margin-top: 20px;
+  margin-bottom: 50px;
+}
