+++ app/components/commitfunds/CommitUnknownUserEstimation.scss
@@ -2,40 +2,6 @@
 
 $spacing-from-container: 22px;
 
-.text-copyable {
-  font-size: 22px;
-}
-
-.initial-link {
-  margin-top: 20px;
-  margin-bottom: 50px;
-}
-
-.info-container {
-  border-radius: 10px;
-  box-shadow: 0 1px 3px 0 rgba(0, 0, 0, .25);  // sass-lint:disable-line no-color-literals
-  background: linear-gradient(145deg, #f2f9eb 0%, #f8f9fa 100%);  // sass-lint:disable-line no-color-literals
-  padding: $spacing-from-container;
-  line-height: 1.6;
-  font-size: $font-size-smallest;
-
-  p {
-    margin-bottom: 12px;
-  }
-
-  .verify-link {
-    margin-left: 30px;
-  }
-}
-
-.gas-section {
-  margin-top: 20px;
-
-  .text-copyable {
-    font-size: $font-size-small;
-  }
-}
-
 .estimation-component {
   padding-top: $spacing-from-container;
   font-size: $font-size-smallest;
